# Replace ContainsRune with IndexByte