After
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/libinjection-go
BenchmarkSQLiDriver
BenchmarkSQLiDriver/sqli
BenchmarkSQLiDriver/sqli-10         	   32134	     36256 ns/op	   49584 B/op	     321 allocs/op
BenchmarkSQLiDriver/folding
BenchmarkSQLiDriver/folding-10      	   12814	     93604 ns/op	   70664 B/op	    1775 allocs/op
BenchmarkSQLiDriver/tokens
BenchmarkSQLiDriver/tokens-10       	    7065	    165434 ns/op	  149562 B/op	    3689 allocs/op
PASS
```

Before
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/libinjection-go
BenchmarkSQLiDriver
BenchmarkSQLiDriver/sqli
BenchmarkSQLiDriver/sqli-10         	   26838	     38006 ns/op	   49584 B/op	     321 allocs/op
BenchmarkSQLiDriver/folding
BenchmarkSQLiDriver/folding-10      	   12643	     94539 ns/op	   70664 B/op	    1775 allocs/op
BenchmarkSQLiDriver/tokens
BenchmarkSQLiDriver/tokens-10       	    7063	    167293 ns/op	  149562 B/op	    3689 allocs/op
PASS
```