@jptosso 2022-09-14
Is there any performance improvement in replacing ContainsRune to BytesIndex?