@anuraaga 2022-09-14
Ah sorry forgot to add the benchmark, added above. A small one